# openapi .json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Define an API",
    "title": "Restful API(connect two databases)",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "http://www.example.com/support",
      "email": "a167829435@gmail.com"
    }
  },
  "paths": {
    "/mysql/signup": {
      "post": {
        "description": "註冊",
        "requestBody": {
          "$ref": "#/components/requestBodies/models.user"
        },
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.user"
                },
                "example": {
                  "email": "test@example.com",
                  "password": ""
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "E-mail or Password error!"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "E-mail already taken!"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "create a new account",
        "tags": [
          "Mysql_User"
        ]
      }
    },
    "/mysql/login": {
      "post": {
        "description": "登入",
        "requestBody": {
          "$ref": "#/components/requestBodies/models.user"
        },
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.JWT"
                },
                "example": {
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImExNjc4Mjk0MzVAZ21haWwuY29tIiwiZXhwIjoxNTc1NTE1NTYyLCJwYXNzd29yZCI6IiQyYSQxMCRtVWFiQ2p4WE9oZEhOWXZzaElGUXB1ZEhTTE1jSGNCU1F3cE1pdVFHN0VIWTZ6SDE1RXd6QyJ9.4h7a50MQmUK56FXTilpo4kQ7LLab47_oaeZ71ulqKrM"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "E-mail or Password error!"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Invaild Password!"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "login",
        "tags": [
          "Mysql_User"
        ]
      }
    },
    "/mssql/signup": {
      "post": {
        "description": "註冊",
        "requestBody": {
          "$ref": "#/components/requestBodies/models.user"
        },
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.user"
                },
                "example": {
                  "email": "test@example.com",
                  "password": ""
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "E-mail or Password error!"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "E-mail already taken!"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "create a new account",
        "tags": [
          "Mssql_User"
        ]
      }
    },
    "/mssql/login": {
      "post": {
        "description": "登入",
        "requestBody": {
          "$ref": "#/components/requestBodies/models.user"
        },
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.JWT"
                },
                "example": {
                  "token": "MCRtVWFiQ2p4WE9oZEhOWXZzaElGUXB1ZEhTTE1jSGNCU1F3cE1pdVFHN0VIWTZ6SDE1RXd6QyJ9.4h7a50MQmUK56FXTilpo4kQ7LLab47_oaeZ71ulqKrM"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "E-mail or Password error!"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Invaild Password!"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "login",
        "tags": [
          "Mssql_User"
        ]
      }
    },
...
    "/v1/join/getall": {
      "get": {
        "description": "取得所有資料",
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.JoinTable"
                },
                "example": {
                  "Key1": "10600",
                  "Key2": "10604240020017",
                  "Key3": "A24WZ",
                  "Number": 100,
                  "Gender": 1,
                  "Birth": 83,
                  "Injury_degree": "3",
                  "Injury_position": 10,
                  "Protection": 1,
                  "Phone": 2,
                  "Person": "B03",
                  "Car": "7",
                  "Action_status": 1,
                  "Qualification": 2,
                  "License": 3,
                  "Drinking": 1,
                  "Hit": 2,
                  "City": "台中",
                  "Position": "國道一號",
                  "Lane": "外側",
                  "Death": "0",
                  "Injured": "2",
                  "Death_exceed": "0",
                  "Weather": "晴",
                  "Light": "日間光線",
                  "Time_year": 2017,
                  "Time_month": "4",
                  "Accident_chinese": "車與車",
                  "Anecdote_chinese": "駕駛人"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "取得關聯表部分資料",
        "tags": [
          "JoinTable"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1"
    }
  ],
  "components": {
    "requestBodies": {
      "models.user": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/models.user"
            },
            "example": {
              "id": "example@gmail.com",
              "password": "a123456"
            }
          }
        },
        "description": "個人資料",
        "required": true
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "models.Error": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "models.JWT": {
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "models.JoinTable": {
        "properties": {
          "accident_chinese": {
            "type": "string"
          },
          "action_status": {
            "type": "integer"
          },
          "anecdote_chinese": {
            "type": "string"
          },
          "birth": {
            "type": "integer"
          },
          "car": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "death": {
            "type": "string"
          },
          "death_exceed": {
            "type": "string"
          },
          "drinking": {
            "type": "integer"
          },
          "gender": {
            "type": "integer"
          },
          "hit": {
            "type": "integer"
          },
          "injured": {
            "type": "string"
          },
          "injury_degree": {
            "type": "string"
          },
          "injury_position": {
            "type": "integer"
          },
          "key1": {
            "type": "string"
          },
          "key2": {
            "type": "string"
          },
          "key3": {
            "type": "string"
          },
          "lane": {
            "type": "string"
          },
          "license": {
            "type": "integer"
          },
          "light": {
            "type": "string"
          },
          "number": {
            "type": "integer"
          },
          "person": {
            "type": "string"
          },
          "phone": {
            "type": "integer"
          },
          "position": {
            "type": "string"
          },
          "protection": {
            "type": "integer"
          },
          "qualification": {
            "type": "integer"
          },
          "time_month": {
            "type": "string"
          },
          "time_year": {
            "type": "integer"
          },
          "weather": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "models.user": {
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "type": "object"
      }
    }
  }
}